Module Examples_without_actions.Internal_example_instantiation

Example instantiation with strings for symbols

type ('a, 'b) grammar = {

name : string;

descr : string;

initial_nt : 'a;

rules : 'b;

}

A named tuple for tagging grammars in a slightly more digestible form than a plain tuple

type nt = string

type tm = string

type sym = (nt, tm) Tjr_simple_earley.Earley_intf.generic_sym

type rule = nt * sym list

val is_nt : string -> bool

Hack to determine nt/tm based on string repr starting with a capital letter

val string_to_sym : string -> (string, string) Tjr_simple_earley.Earley_intf.generic_sym

val make_rule : 'a -> string list -> 'a * (string, string) Tjr_simple_earley.Earley_intf.generic_sym list

val _1 : 'a -> 'a list

val _2 : ('a * 'a) -> 'a list

val _3 : ('a * 'a * 'a) -> 'a list

val _E : string

val _S : string

val a : 'a -> 'a

val eps : string

val one : string

val x : string

val grammar : name:string -> descr:string -> initial_nt:'a -> rules:'b -> ('a, 'b) grammar

val example_grammars : (string, (string * (string, string) Tjr_simple_earley.Earley_intf.generic_sym list) list) grammar list

module Export : sig ... end