Module Tjr_simple_earley_examples__Actions

Apply actions to the result of an Earley parse.

module Internal : functor (T : sig ... end) -> sig ... end

val apply_actions : is_nt:('sym -> bool) -> dest_nt:('sym -> 'nt) -> dest_tm:('sym -> 'tm) -> get_rhss:(nt:'nt -> ('sym list * ('uni_val list -> 'uni_val)) list) -> cut:(int -> ('sym * 'sym list) -> int -> int list) -> apply_tm:(tm:'tm -> i:int -> j:int -> 'uni_val option) -> nt:'nt -> i:int -> j:int -> (int * 'nt * int) list -> 'uni_val option

Apply actions, given various pieces of information about the parse (see Internal for further comments).