Module Tjr_simple_earley_examples__Actions.Internal

Parameters

T : sig ... end

Signature

val apply_actions : is_nt:(T.sym -> bool) -> dest_nt:(T.sym -> T.nt) -> dest_tm:(T.sym -> T.tm) -> get_rhss:(nt:T.nt -> (T.sym list * (T.uni_val list -> T.uni_val)) list) -> cut:(int -> (T.sym * T.sym list) -> int -> int list) -> apply_tm:(tm:T.tm -> i:int -> j:int -> T.uni_val option) -> nt:T.nt -> i:int -> j:int -> (int * T.nt * int) list -> T.uni_val option

Apply actions for a given symbol, between two indices.

NOTES

argument apply_tm takes a tm and a span (i,j) and returns at most one result

for/tm: see comment in code

for_nt/false case: uses backtracking to iterate over the possibilities

for_nt/false case: we iterate over the cut options till we find a possible rhs, and then we apply the action

for_syms cases: we treat and sym specially